+++ include/open_atmos/mechanism_configuration/parser_types.hpp
@@ -65,5 +65,94 @@ namespace open_atmos
           open_atmos::types::Reactions& reactions) override;
     };
 
+    class CondensedPhasePhotolysisParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class SimpolPhaseTransferParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class EmissionParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class PhotolysisParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class FirstOrderLossParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class AqueousEquilibriumParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class WetDepositionParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class HenrysLawParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
+
+    class ArrheniusParser : public IReactionParser
+    {
+     public:
+      ConfigParseStatus parse(
+          const YAML::Node& object,
+          const std::vector<types::Species>& existing_species,
+          const std::vector<types::Phase>& existing_phases,
+          open_atmos::types::Reactions& reactions) override;
+    };
   }  // namespace mechanism_configuration
 }  // namespace open_atmos
