+++ src/CMakeLists.txt
@@ -18,7 +18,6 @@ target_include_directories(mechanism_configuration
     $<INSTALL_INTERFACE:include>
 )
 
-
 target_sources(mechanism_configuration
   PRIVATE
     parser.cpp
@@ -29,6 +28,15 @@ target_sources(mechanism_configuration
     branched_parser.cpp
     tunneling_parser.cpp
     surface_parser.cpp
+    photolysis_parser.cpp
+    condensed_phase_photolysis_parser.cpp
+    emission_parser.cpp
+    first_order_loss_parser.cpp
+    simpol_phase_transfer_parser.cpp
+    aqueous_equilibrium_parser.cpp
+    wet_deposition_parser.cpp
+    henrys_law_parser.cpp
+    arrhenius_parser.cpp
 )
 
 target_link_libraries(mechanism_configuration
